✨Fix: LD-156 코멘트 수정
- LD-156 코멘트 수정
