+++ Projects/LastDance/LastDance/Sources/Domain/Model/MockExhibitionData.swift
@@ -1,0 +1,16 @@
+//
+//  MockExhibitionData.swift
+//  LastDance
+//
+//  Created by 광로 on 10/21/25.
+//
+
+import Foundation
+
+struct MockExhibitionData: Identifiable {
+    let id: String
+    let title: String
+    let coverImageName: String
+    let reactionCount: Int
+}
+
